first part refactor
