+++ hawaii-cache/src/main/java/org/hawaiiframework/cache/redis/HawaiiRedisCacheBuilder.java
@@ -1,0 +1,8 @@
+package org.hawaiiframework.cache.redis;
+
+public class HawaiiRedisCacheBuilder {
+
+
+
+
+}
